refactor: Corrigindo estrutura de modal de edição e exclusão
