+++ src/components/modals/index.js
@@ -1,0 +1,3 @@
+export * from "./edit-modal";
+export * from "./exclude-student";
+
